oracle: type ambiguous computed NUMBER as decimal (#844)
A query whose result is a computed NUMBER with a fractional value (e.g. a
division like .actor_id / 8, or a native AVG) failed at scan time on Oracle:
sq typed the bare NUMBER result column as int64, but go-ora returns the
fractional value as a string, so "7.25" could not parse into int64. Native
sq sql was affected too.

Oracle reports no usable precision or scale for a computed NUMBER: the TTC
describe carries precision=0, scale=0 for COUNT(*), MAX/MIN, division, AVG,
SUM, and integer arithmetic alike, so the result is genuinely ambiguous
between integer and fractional. Confirmed against both the go-ora wire
describe and Oracle's ojdbc JDBC driver, which maps every such NUMBER to
BigDecimal. Classify the floating-scale form as kind.Decimal, which is exact
and scans any value without error, for both SLQ and native SQL.

count(), count_unique(), and rownum() are integer-valued but share the same
ambiguous form. Pin them back to kind.Int via Renderer.FunctionResultKinds,
which RecordMeta applies from the rendering context (the mechanism that
SQLite and rqlite already use for sum()).

Consequence: an integer-valued computed result on Oracle (max/min, an
integer literal, integer arithmetic, or COUNT(*) via sq sql) is now a
decimal, rendered as a quoted string in JSON output. Operand-aware integer
typing for SLQ is tracked in #838; decimal string-vs-number rendering in #846.

+++ drivers/oracle/internal_test.go
@@ -156,10 +156,12 @@ func TestRefineBareNumberKind(t *testing.T) {
 		ok               bool
 		want             kind.Kind
 	}{
-		// COUNT(*), SUM, integer literals — Oracle reports (38, 255).
-		{"floating_38_255", 38, oracleScaleFloating, true, kind.Int},
-		// Same shape, different precision — still floating.
-		{"floating_0_255", 0, oracleScaleFloating, true, kind.Int},
+		// The floating-scale form (scale 255) is ambiguous: COUNT(*), SUM, AVG,
+		// integer literals, and division all report it. It maps to kind.Decimal
+		// so fractional values (e.g. division) don't crash an int64 scan (#844).
+		{"floating_38_255", 38, oracleScaleFloating, true, kind.Decimal},
+		// Same shape, different precision — still floating, still decimal.
+		{"floating_0_255", 0, oracleScaleFloating, true, kind.Decimal},
 		// NUMBER(p,0) with p in [1..19] is int range.
 		{"int_5_0", 5, 0, true, kind.Int},
 		{"int_19_0", 19, 0, true, kind.Int},
